move to replp.Operators._ namespace
